cositas basicas de la implementacion

+++ proyecto_final/src/main/java/com/unimag/edu/proyecto_final/domine/entities/Baggage.java
@@ -20,7 +20,7 @@ public class Baggage {
     @JoinColumn(name = "ticket_id")
     private Ticket ticket;
     private Double weight;
-    private BigDecimal fee;
+    private Double fee;
     private String tagCode;
 
 }
